Object Classifier/requirements.txt
tensorflow
numpy
requests
Pillow